JNM: Extracted JavaNativeClassConverter
